+++ packages/cli/src/commands/migrate/tasks/config-lint.ts
@@ -1,11 +1,12 @@
-import { resolve } from 'path';
+import { resolve, join } from 'path';
 import { ESLint } from 'eslint';
-import { existsSync, outputFileSync } from 'fs-extra';
+import { glob } from 'glob';
+import { outputFileSync } from 'fs-extra';
 import defaultConfig from '../../../configs/default-eslint';
+import { gitAddIfInRepo } from '../../../utils';
 import type { ListrTask } from 'listr2';
 import type { MigrateCommandContext, MigrateCommandOptions } from '../../../types';
 
-const CONFIG_FILENAME = '.eslintrc.js';
 const REHEARSAL_CONFIG_FILENAME = '.rehearsal-eslintrc.js';
 const REHEARSAL_CONFIG_RELATIVE_PATH = `./${REHEARSAL_CONFIG_FILENAME}`;
 
@@ -14,22 +15,25 @@ export async function lintConfigTask(options: MigrateCommandOptions): Promise<Li
     title: 'Create eslint config',
     enabled: (ctx: MigrateCommandContext): boolean => !ctx.skip,
     task: async (_ctx: MigrateCommandContext, task): Promise<void> => {
+      // glob against the following file extension pattern js,yml,json,yaml and return the first match
+      const configPath = glob.sync(join(options.basePath, '.eslintrc.{js,yml,json,yaml}'))[0];
+
       if (_ctx.userConfig?.hasLintSetup) {
         task.output = `Create .eslintrc.js from config`;
         await _ctx.userConfig.lintSetup();
-      }
-
-      createRehearsalConfig(options.basePath);
-
-      const configPath = resolve(options.basePath, CONFIG_FILENAME);
-
-      if (configExists(configPath)) {
-        task.output = `${configPath} already exists, extending Rehearsal default typescript-related config`;
-        task.title = `Update eslintrc.js`;
-        await extendsRehearsalInCurrentConfig(configPath, REHEARSAL_CONFIG_RELATIVE_PATH);
       } else {
-        task.output = `Create .eslintrc.js, extending Rehearsal default typescript-related config`;
-        extendsRehearsalInNewConfig(configPath, REHEARSAL_CONFIG_RELATIVE_PATH);
+        // only run the default process with no custom config provided
+        // create .rehearsal-eslintrc.js
+        createRehearsalConfig(options.basePath);
+
+        if (configPath) {
+          task.output = `${configPath} already exists, extending Rehearsal default typescript-related config`;
+          task.title = `Update eslintrc.js`;
+          await extendsRehearsalInCurrentConfig(configPath, REHEARSAL_CONFIG_RELATIVE_PATH);
+        } else {
+          task.output = `Create .eslintrc.js, extending Rehearsal default typescript-related config`;
+          extendsRehearsalInNewConfig(options.basePath, REHEARSAL_CONFIG_RELATIVE_PATH);
+        }
       }
     },
   };
@@ -39,6 +43,7 @@ function createRehearsalConfig(basePath: string): void {
   const rehearsalConfigStr = 'module.exports = ' + JSON.stringify(defaultConfig, null, 2);
   const rehearsalConfigPath = resolve(basePath, REHEARSAL_CONFIG_FILENAME);
   outputFileSync(rehearsalConfigPath, rehearsalConfigStr);
+  gitAddIfInRepo(rehearsalConfigPath); // stage '.rehearsal-eslintrc.js'; if in a git repo
 }
 
 async function extendsRehearsalInCurrentConfig(
@@ -58,9 +63,10 @@ async function extendsRehearsalInCurrentConfig(
 }
 
 async function extendsRehearsalInNewConfig(
-  configPath: string,
+  basePath: string,
   rehearsalConfigRelativePath: string
 ): Promise<void> {
+  const configPath = resolve(basePath, '.eslintrc.js');
   const configStr = `
   module.exports = {
     extends: ['${rehearsalConfigRelativePath}']
@@ -73,14 +79,11 @@ async function writeLintConfig(path: string, config: string): Promise<void> {
   outputFileSync(path, config);
   const formattedConfig = await formatLintConfig(config, path);
   outputFileSync(path, formattedConfig);
+  gitAddIfInRepo(path); // stage .eslintrc.js if in a git repo
 }
 
 async function formatLintConfig(configStr: string, filePath: string): Promise<string | undefined> {
   const eslint = new ESLint({ fix: true, useEslintrc: true });
   const [report] = await eslint.lintText(configStr, { filePath: filePath });
   return report.output;
-}
-
-function configExists(configPath: string): boolean {
-  return existsSync(configPath);
 }