+++ packages/cli/test/commands/migrate.test.ts
@@ -1,8 +1,8 @@
 import { resolve } from 'path';
-import { readdirSync, readJSONSync, writeJSONSync } from 'fs-extra';
+import { readdirSync, readJSONSync, writeFileSync, writeJSONSync } from 'fs-extra';
 import { setGracefulCleanup } from 'tmp';
-import { beforeAll, beforeEach, describe, expect, test } from 'vitest';
-import { type SimpleGit, type SimpleGitOptions, simpleGit } from 'simple-git';
+import { beforeEach, describe, expect, test } from 'vitest';
+import { type SimpleGit, simpleGit, type SimpleGitOptions } from 'simple-git';
 
 import { runBin, prepareTmpDir } from '../test-helpers';
 import { REQUIRED_DEPENDENCIES } from '../../src/commands/migrate/tasks/dependency-install';
@@ -92,7 +92,7 @@ describe('migrate - check repo status', async () => {
 describe('migrate - initialization', async () => {
   let basePath = '';
 
-  beforeAll(() => {
+  beforeEach(() => {
     basePath = prepareTmpDir('basic');
   });
 
@@ -101,7 +101,10 @@ describe('migrate - initialization', async () => {
       cwd: basePath,
     });
 
-    expect(result.stdout).toMatchSnapshot();
+    // remove first line that contains rehearsal version
+    const stdoutWithoutFirstLine = result.stdout.split('\n').slice(1).join('\n');
+
+    expect(stdoutWithoutFirstLine).toMatchSnapshot();
   });
 
   // TODO: add tests for other cases
@@ -111,7 +114,7 @@ describe('migrate - initialization', async () => {
 describe('migrate - install dependencies', async () => {
   let basePath = '';
 
-  beforeAll(() => {
+  beforeEach(() => {
     basePath = prepareTmpDir('initialization');
   });
 
@@ -126,12 +129,32 @@ describe('migrate - install dependencies', async () => {
     expect(result.stdout).toContain('Install dependencies');
     expect(Object.keys(devDeps).sort()).toEqual(REQUIRED_DEPENDENCIES.sort());
   });
+
+  test('Install custom dependencies', async () => {
+    createUserConfig(basePath, {
+      migrate: {
+        install: {
+          dependencies: [],
+          devDependencies: ['fs-extra'],
+        },
+      },
+    });
+    const result = await runBin('migrate', ['-u', 'rehearsal-config.json'], {
+      cwd: basePath,
+    });
+
+    const packageJson = readJSONSync(resolve(basePath, 'package.json'));
+    const devDeps = packageJson.devDependencies;
+
+    expect(result.stdout).toContain('Install dependencies from config');
+    expect(devDeps).toHaveProperty('fs-extra');
+  });
 });
 
 describe('migrate - generate tsconfig', async () => {
   let basePath = '';
 
-  beforeAll(() => {
+  beforeEach(() => {
     basePath = prepareTmpDir('initialization');
   });
 
@@ -144,8 +167,33 @@ describe('migrate - generate tsconfig', async () => {
     expect(readdirSync(basePath)).toContain('tsconfig.json');
   });
 
+  test('stage tsconfig if in git repo', async () => {
+    // simulate clean git project
+    const git: SimpleGit = simpleGit({
+      baseDir: basePath,
+    } as Partial<SimpleGitOptions>);
+    // Init git, add and commit existed files, to make it a clean state
+    await git.init();
+    await git.add(resolve(basePath, 'package.json'));
+    // GH CI would require git name and email
+    await git.addConfig('user.name', 'tester');
+    await git.addConfig('user.email', 'tester@tester.com');
+    await git.commit('foo');
+
+    const result = await runBin('migrate', [], {
+      cwd: basePath,
+    });
+
+    const gitStatus = await git.status();
+    expect(gitStatus.staged).toContain('tsconfig.json');
+
+    expect(result.stdout).toContain('Create tsconfig.json');
+    expect(readdirSync(basePath)).toContain('tsconfig.json');
+  });
+
   test('On tsconfig exists, ensure strict mode', async () => {
-    // tsconfig already created from previous test
+    const oldConfig = { compilerOptions: { strict: false } };
+    writeJSONSync(resolve(basePath, 'tsconfig.json'), oldConfig);
     const result = await runBin('migrate', [], {
       cwd: basePath,
     });
@@ -157,8 +205,7 @@ describe('migrate - generate tsconfig', async () => {
     expect(tsConfig.compilerOptions.strict).toBeTruthy;
   });
 
-  test('runBin custom ts config command with user config provided', async () => {
-    basePath = prepareTmpDir('initialization');
+  test('Custom ts config command with user config provided', async () => {
     createUserConfig(basePath, {
       migrate: {
         setup: {
@@ -182,23 +229,26 @@ describe('migrate - JS to TS conversion', async () => {
     basePath = prepareTmpDir('basic');
   });
 
-  test('able to migrate from default index.js', async () => {
+  test('able to migrate from default all files .js in root', async () => {
     const result = await runBin('migrate', [], {
       cwd: basePath,
     });
 
     // Test summary message
-    expect(result.stdout).toContain(`2 JS files converted to TS`);
+    expect(result.stdout).toContain(`3 JS files converted to TS`);
 
     expect(readdirSync(basePath)).toContain('index.ts');
     expect(readdirSync(basePath)).toContain('foo.ts');
+    expect(readdirSync(basePath)).toContain('depends-on-foo.ts');
 
     expect(readdirSync(basePath)).not.toContain('index.js');
     expect(readdirSync(basePath)).not.toContain('foo.js');
+    expect(readdirSync(basePath)).not.toContain('depends-on-foo.js');
 
     const config = readJSONSync(resolve(basePath, 'tsconfig.json'));
     expect(config.include).toContain('index.ts');
     expect(config.include).toContain('foo.ts');
+    expect(config.include).toContain('depends-on-foo.ts');
   });
 
   test('able to migrate from specific entrypoint', async () => {
@@ -236,6 +286,35 @@ describe('migrate - JS to TS conversion', async () => {
     expect(readdirSync(reportPath)).toContain('migrate-report.sarif');
   });
 
+  test('stage report if in a git repo', async () => {
+    // simulate clean git project
+    const git: SimpleGit = simpleGit({
+      baseDir: basePath,
+    } as Partial<SimpleGitOptions>);
+    // Init git, add and commit existed files, to make it a clean state
+    await git.init();
+    await git.add(readdirSync(basePath));
+    // GH CI would require git name and email
+    await git.addConfig('user.name', 'tester');
+    await git.addConfig('user.email', 'tester@tester.com');
+    await git.commit('foo');
+
+    await runBin('migrate', [], {
+      cwd: basePath,
+    });
+
+    const reportPath = resolve(basePath, '.rehearsal');
+    expect(readdirSync(reportPath)).toContain('migrate-report.sarif');
+
+    const gitStatus = await git.status();
+    expect(gitStatus.staged).toStrictEqual([
+      '.eslintrc.js',
+      '.rehearsal-eslintrc.js',
+      '.rehearsal/migrate-report.sarif',
+      'tsconfig.json',
+    ]);
+  });
+
   test('Generate report in different formats with -f flag', async () => {
     await runBin('migrate', ['-f', 'json,md,sarif,foo'], {
       cwd: basePath,
@@ -253,8 +332,68 @@ describe('migrate - JS to TS conversion', async () => {
 describe('migrate - generate eslint config', async () => {
   let basePath = '';
 
-  beforeAll(() => {
+  beforeEach(() => {
     basePath = prepareTmpDir('initialization');
+  });
+
+  test('create .eslintrc.js if not existed', async () => {
+    const result = await runBin('migrate', [], {
+      cwd: basePath,
+    });
+
+    expect(result.stdout).toContain(
+      'Create .eslintrc.js, extending Rehearsal default typescript-related config'
+    );
+    expect(readdirSync(basePath)).toContain('.eslintrc.js');
+    expect(readdirSync(basePath)).toContain('.rehearsal-eslintrc.js');
+
+    /* eslint-disable-next-line @typescript-eslint/no-var-requires */
+    const eslintConfig = require(resolve(basePath, '.eslintrc.js'));
+    expect(eslintConfig.extends).toStrictEqual(['./.rehearsal-eslintrc.js']);
+  });
+
+  test('extends .eslintrc.js if existed', async () => {
+    const oldConfig = `
+    module.exports = {extends: ['foo']};
+  `;
+    writeFileSync(resolve(basePath, '.eslintrc.js'), oldConfig);
+
+    const result = await runBin('migrate', [], {
+      cwd: basePath,
+    });
+
+    expect(result.stdout).toContain('extending Rehearsal default typescript-related config');
+    expect(readdirSync(basePath)).toContain('.eslintrc.js');
+    expect(readdirSync(basePath)).toContain('.rehearsal-eslintrc.js');
+
+    /* eslint-disable-next-line @typescript-eslint/no-var-requires */
+    const newConfig = require(resolve(basePath, '.eslintrc.js'));
+    expect(newConfig.extends).toStrictEqual(['foo', './.rehearsal-eslintrc.js']);
+  });
+
+  test('stage lint configs if in git repo', async () => {
+    // simulate clean git project
+    const git: SimpleGit = simpleGit({
+      baseDir: basePath,
+    } as Partial<SimpleGitOptions>);
+    // Init git, add and commit existed files, to make it a clean state
+    await git.init();
+    await git.add(resolve(basePath, 'package.json'));
+    // GH CI would require git name and email
+    await git.addConfig('user.name', 'tester');
+    await git.addConfig('user.email', 'tester@tester.com');
+    await git.commit('foo');
+
+    await runBin('migrate', [], {
+      cwd: basePath,
+    });
+
+    expect(readdirSync(basePath)).toContain('.eslintrc.js');
+    expect(readdirSync(basePath)).toContain('.rehearsal-eslintrc.js');
+
+    const gitStatus = await git.status();
+    expect(gitStatus.staged).toContain('.eslintrc.js');
+    expect(gitStatus.staged).toContain('.rehearsal-eslintrc.js');
   });
 
   test('Run custom lint config command with user config provided', async () => {
@@ -279,7 +418,7 @@ describe('migrate - generate eslint config', async () => {
 describe('migrate - handle custom basePath', async () => {
   let basePath = '';
 
-  beforeAll(() => {
+  beforeEach(() => {
     basePath = prepareTmpDir('custom_basepath');
   });
 
@@ -303,7 +442,7 @@ describe('migrate - handle custom basePath', async () => {
 describe('migrate - new scripts for TS', async () => {
   let basePath = '';
 
-  beforeAll(() => {
+  beforeEach(() => {
     basePath = prepareTmpDir('basic');
   });