+++ packages/cli/test/helpers/utils.test.ts
@@ -11,6 +11,8 @@ import {
   getPathToBinary,
   isPnpmManager,
   isYarnManager,
+  isBinExisted,
+  getManagerBinPath,
   normalizeVersionString,
   sleep,
   timestamp,
@@ -71,6 +73,24 @@ describe('utils', () => {
     expect(isPnpm).equal(true);
   });
 
+  test('isBinExisted()', () => {
+    expect(isBinExisted('ls')).equal(true);
+    expect(isBinExisted('this-should-not-exist')).equal(false);
+  });
+
+  test('getManagerBinPath()', () => {
+    // TODO: Add test scenarios if volta exists
+    // Haven't came up with a quick and good way to do it
+    const npmBinPath = execa.sync('which', ['npm']).stdout;
+    expect(getManagerBinPath('npm', false)).toBe(npmBinPath);
+
+    const yarnBinPath = execa.sync('which', ['yarn']).stdout;
+    expect(getManagerBinPath('yarn', false)).toBe(yarnBinPath);
+
+    const pnpmBinPath = execa.sync('which', ['pnpm']).stdout;
+    expect(getManagerBinPath('pnpm', false)).toBe(pnpmBinPath);
+  });
+
   test('getModuleManager()', () => {
     const manager = getModuleManager();
 
@@ -85,11 +105,15 @@ describe('utils', () => {
   });
 
   test('getLatestTSVersion()', async () => {
-    const latestVersionTagged = compare(await getLatestTSVersion('beta'), '1.0.0', '>');
-    const latestVersion = compare(await getLatestTSVersion('latestBeta'), '1.0.0', '>');
+    const betaVersionTagged = compare(await getLatestTSVersion('beta'), '4.0.0', '>');
+    const latestBetaVersion = compare(await getLatestTSVersion('latestBeta'), '4.0.0', '>');
+    const latestVersion = compare(await getLatestTSVersion('latest'), '4.0.0', '>');
+    const rcVersion = compare(await getLatestTSVersion('rc'), '4.0.0', '>');
 
-    expect(latestVersionTagged).toBeTruthy();
+    expect(betaVersionTagged).toBeTruthy();
     expect(latestVersion).toBeTruthy();
+    expect(latestBetaVersion).toBeTruthy();
+    expect(rcVersion).toBeTruthy();
   });
 
   test('getLockfilePath()', () => {