+++ packages/migration-graph-ember/src/types.ts
@@ -1,5 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { type PackageContainer } from '@rehearsal/migration-graph-shared';
+import type { Package, PackageContainer } from '@rehearsal/migration-graph-shared';
+import type { EmberAddonPackage } from './entities/ember-addon-package';
+import type { EmberAppPackage } from './entities/ember-app-package';
 
 export type EmberPackageContainer = {
   getInternalPackages?: (...args: any) => unknown;
@@ -9,3 +11,5 @@ export type EmberPackageContainer = {
   getAddonPackages?: (...args: any) => unknown;
   getRootPackage?: (...args: any) => unknown;
 } & PackageContainer;
+
+export type EmberProjectPackage = Package | EmberAddonPackage | EmberAppPackage;