+++ packages/migration-graph-shared/test/entities/package.test.ts
@@ -48,11 +48,26 @@ describe('Unit | Entities | Package', function () {
     });
   });
 
-  describe('includ/exclude patterns', () => {
+  describe('include/exclude patterns', () => {
     test('defaults', () => {
       const p = new Package(pathToPackage);
-      expect(p.excludePatterns).toStrictEqual(new Set(['dist', 'test', 'tests']));
-      expect(p.includePatterns).toStrictEqual(new Set(['index.js']));
+      expect(p.excludePatterns).toStrictEqual(
+        new Set([
+          '.yarn',
+          'dist',
+          'test',
+          'tests',
+          '.eslintrc.*',
+          '.babelrc.*',
+          'babel.config.*',
+          'Brocfile.js',
+          'prettier.config.*',
+          'karma.config.*',
+          'webpack.config.js',
+          'vite.config.ts',
+        ])
+      );
+      expect(p.includePatterns).toStrictEqual(new Set(['.']));
     });
 
     test('options.excludePatterns ', () => {
@@ -70,11 +85,22 @@ describe('Unit | Entities | Package', function () {
     test('addExcludePattern', () => {
       const p = new Package(pathToPackage);
       p.addExcludePattern('test-packages');
-      expect(p.excludePatterns).toStrictEqual(new Set(['dist', 'test', 'tests', 'test-packages']));
-
-      p.addExcludePattern('file1', 'file2');
       expect(p.excludePatterns).toStrictEqual(
-        new Set(['dist', 'test', 'tests', 'test-packages', 'file1', 'file2'])
+        new Set([
+          '.yarn',
+          'dist',
+          'test',
+          'tests',
+          '.eslintrc.*',
+          '.babelrc.*',
+          'babel.config.*',
+          'Brocfile.js',
+          'prettier.config.*',
+          'karma.config.*',
+          'webpack.config.js',
+          'vite.config.ts',
+          'test-packages',
+        ])
       );
     });
 
@@ -82,10 +108,10 @@ describe('Unit | Entities | Package', function () {
       const p = new Package(pathToPackage);
       p.addIncludePattern('foo.js');
       expect(p.includePatterns.has('foo.js')).toBeTruthy();
-      expect(p.includePatterns).toStrictEqual(new Set(['index.js', 'foo.js']));
+      expect(p.includePatterns).toStrictEqual(new Set(['.', 'foo.js']));
 
       p.addIncludePattern('file1', 'file2');
-      expect(p.includePatterns).toStrictEqual(new Set(['index.js', 'foo.js', 'file1', 'file2']));
+      expect(p.includePatterns).toStrictEqual(new Set(['.', 'foo.js', 'file1', 'file2']));
     });
   });