+++ CHANGELOG.md
@@ -1,3 +1,86 @@
+## <small>1.0.3-beta (2023-01-21)</small>
+
+* chore: make sure the repo is on master with all latest changes before starting release ([cfcb5a1](https://github.com/rehearsal-js/rehearsal-js/commit/cfcb5a1))
+* chore: release script (#647) ([87414d9](https://github.com/rehearsal-js/rehearsal-js/commit/87414d9)), closes [#647](https://github.com/rehearsal-js/rehearsal-js/issues/647)
+* fix: Correct peerDependencies typescript version (#645) ([7704b46](https://github.com/rehearsal-js/rehearsal-js/commit/7704b46)), closes [#645](https://github.com/rehearsal-js/rehearsal-js/issues/645)
+* fix: package manager version conflict with volta (#646) ([0f252bc](https://github.com/rehearsal-js/rehearsal-js/commit/0f252bc)), closes [#646](https://github.com/rehearsal-js/rehearsal-js/issues/646)
+* fix: release script (#650) ([6e64d3b](https://github.com/rehearsal-js/rehearsal-js/commit/6e64d3b)), closes [#650](https://github.com/rehearsal-js/rehearsal-js/issues/650)
+* fix: remove test's rehearsal version dependency (#649) ([4b02faf](https://github.com/rehearsal-js/rehearsal-js/commit/4b02faf)), closes [#649](https://github.com/rehearsal-js/rehearsal-js/issues/649)
+
+
+
+## v1.0.2-beta
+
+* fix: bundle typescript and bump (#634) ([2f7dd21](https://github.com/rehearsal-js/rehearsal-js/commit/2f7dd21)), closes [#634](https://github.com/rehearsal-js/rehearsal-js/issues/634)
+* fix: issue with custom config (#625) ([a28986c](https://github.com/rehearsal-js/rehearsal-js/commit/a28986c)), closes [#625](https://github.com/rehearsal-js/rehearsal-js/issues/625)
+* fix(migration-graph-ember): addon; ignore app and index.js (#627) ([2746e64](https://github.com/rehearsal-js/rehearsal-js/commit/2746e64)), closes [#627](https://github.com/rehearsal-js/rehearsal-js/issues/627)
+* feat: commit all the config files during migration (#629) ([96e4384](https://github.com/rehearsal-js/rehearsal-js/commit/96e4384)), closes [#629](https://github.com/rehearsal-js/rehearsal-js/issues/629)
+* feat(migration-graph): ignore more files (#626) ([263c37b](https://github.com/rehearsal-js/rehearsal-js/commit/263c37b)), closes [#626](https://github.com/rehearsal-js/rehearsal-js/issues/626)
+* chore: refactor to use Package.path (#618) ([2697170](https://github.com/rehearsal-js/rehearsal-js/commit/2697170)), closes [#618](https://github.com/rehearsal-js/rehearsal-js/issues/618)
+* chore: remove `src` from tarballs (#617) ([8fe2717](https://github.com/rehearsal-js/rehearsal-js/commit/8fe2717)), closes [#617](https://github.com/rehearsal-js/rehearsal-js/issues/617)
+* chore(migration-graph-ember): tidy types (#619) ([7f05b8a](https://github.com/rehearsal-js/rehearsal-js/commit/7f05b8a)), closes [#619](https://github.com/rehearsal-js/rehearsal-js/issues/619)
+* Update README.md ([322a8bf](https://github.com/rehearsal-js/rehearsal-js/commit/322a8bf))
+
+
+
+## v1.0.1-beta
+
+* chore: test for loose files in a package (#615) ([0705760](https://github.com/rehearsal-js/rehearsal-js/commit/0705760)), closes [#615](https://github.com/rehearsal-js/rehearsal-js/issues/615)
+* chore(migration-graph): improve workspace test with out-of-scope file (#616) ([af04522](https://github.com/rehearsal-js/rehearsal-js/commit/af04522)), closes [#616](https://github.com/rehearsal-js/rehearsal-js/issues/616)
+
+
+
+##  v1.0.0-beta
+
+* BREAKING CHANGE(migration-graph): include all files in package dir by default (#595) ([38257d6](https://github.com/rehearsal-js/rehearsal-js/commit/38257d6)), closes [#595](https://github.com/rehearsal-js/rehearsal-js/issues/595)
+* refactor: add fixed error count in report, revise cli output (#599) ([cdb843a](https://github.com/rehearsal-js/rehearsal-js/commit/cdb843a)), closes [#599](https://github.com/rehearsal-js/rehearsal-js/issues/599)
+* refactor: break down migrate tasks (#541) ([23dbafe](https://github.com/rehearsal-js/rehearsal-js/commit/23dbafe)), closes [#541](https://github.com/rehearsal-js/rehearsal-js/issues/541)
+* refactor: migrate as pull based gen and itr (#612) ([1c7e72d](https://github.com/rehearsal-js/rehearsal-js/commit/1c7e72d)), closes [#612](https://github.com/rehearsal-js/rehearsal-js/issues/612)
+* refactor: remove plugins/data (#600) ([67e9d91](https://github.com/rehearsal-js/rehearsal-js/commit/67e9d91)), closes [#600](https://github.com/rehearsal-js/rehearsal-js/issues/600)
+* refactor: replace @ts-ignore with @ts-expect-error (#504) ([9effaa7](https://github.com/rehearsal-js/rehearsal-js/commit/9effaa7)), closes [#504](https://github.com/rehearsal-js/rehearsal-js/issues/504)
+* refactor: replace codeFixAction add with insert (#528) ([a744f66](https://github.com/rehearsal-js/rehearsal-js/commit/a744f66)), closes [#528](https://github.com/rehearsal-js/rehearsal-js/issues/528)
+* refactor: separate diagnostic fix from check to fix line numbers (#556) ([73a5b96](https://github.com/rehearsal-js/rehearsal-js/commit/73a5b96)), closes [#556](https://github.com/rehearsal-js/rehearsal-js/issues/556)
+* fix: add @rehearsal comment before literal template for any issue inside it (#547) ([7198bd2](https://github.com/rehearsal-js/rehearsal-js/commit/7198bd2)), closes [#547](https://github.com/rehearsal-js/rehearsal-js/issues/547)
+* fix: add TS related scripts in package.json (#503) ([5603708](https://github.com/rehearsal-js/rehearsal-js/commit/5603708)), closes [#503](https://github.com/rehearsal-js/rehearsal-js/issues/503)
+* fix: cleanup cli output (#538) ([d060751](https://github.com/rehearsal-js/rehearsal-js/commit/d060751)), closes [#538](https://github.com/rehearsal-js/rehearsal-js/issues/538)
+* fix: control-c during git mv (#583) ([2590606](https://github.com/rehearsal-js/rehearsal-js/commit/2590606)), closes [#583](https://github.com/rehearsal-js/rehearsal-js/issues/583)
+* fix: correct line and column numbers (#522) ([9494a93](https://github.com/rehearsal-js/rehearsal-js/commit/9494a93)), closes [#522](https://github.com/rehearsal-js/rehearsal-js/issues/522)
+* fix: correct sarif kind and level (#527) ([7e4a80f](https://github.com/rehearsal-js/rehearsal-js/commit/7e4a80f)), closes [#527](https://github.com/rehearsal-js/rehearsal-js/issues/527)
+* fix: detailed console message during JS->TS conversion (#502) ([5189d59](https://github.com/rehearsal-js/rehearsal-js/commit/5189d59)), closes [#502](https://github.com/rehearsal-js/rehearsal-js/issues/502)
+* fix: diagnostic ordering (#524) ([1f97b75](https://github.com/rehearsal-js/rehearsal-js/commit/1f97b75)), closes [#524](https://github.com/rehearsal-js/rehearsal-js/issues/524)
+* fix: display version for migrate and upgrade, issue#484 (#499) ([10965fb](https://github.com/rehearsal-js/rehearsal-js/commit/10965fb)), closes [issue#484](https://github.com/issue/issues/484) [#499](https://github.com/rehearsal-js/rehearsal-js/issues/499)
+* fix: file extension name replacement bug (#572) ([470e30b](https://github.com/rehearsal-js/rehearsal-js/commit/470e30b)), closes [#572](https://github.com/rehearsal-js/rehearsal-js/issues/572)
+* fix: format extensions (#539) ([4681545](https://github.com/rehearsal-js/rehearsal-js/commit/4681545)), closes [#539](https://github.com/rehearsal-js/rehearsal-js/issues/539)
+* fix: format flags (#537) ([64f6acb](https://github.com/rehearsal-js/rehearsal-js/commit/64f6acb)), closes [#537](https://github.com/rehearsal-js/rehearsal-js/issues/537)
+* fix: launch editor issue && show summary message in interactive mode (#577) ([59eb327](https://github.com/rehearsal-js/rehearsal-js/commit/59eb327)), closes [#577](https://github.com/rehearsal-js/rehearsal-js/issues/577)
+* fix: migrate cli should stop on failure (#501) ([32a5944](https://github.com/rehearsal-js/rehearsal-js/commit/32a5944)), closes [#501](https://github.com/rehearsal-js/rehearsal-js/issues/501)
+* fix: migrate cmd handle listr async tasks (#573) ([e5d448e](https://github.com/rehearsal-js/rehearsal-js/commit/e5d448e)), closes [#573](https://github.com/rehearsal-js/rehearsal-js/issues/573)
+* fix: package scope (#596) ([a2b3522](https://github.com/rehearsal-js/rehearsal-js/commit/a2b3522)), closes [#596](https://github.com/rehearsal-js/rehearsal-js/issues/596)
+* fix: perf when crawling deps (#485) ([8df66cb](https://github.com/rehearsal-js/rehearsal-js/commit/8df66cb)), closes [#485](https://github.com/rehearsal-js/rehearsal-js/issues/485)
+* fix: remove property bag (#513) ([49a9b19](https://github.com/rehearsal-js/rehearsal-js/commit/49a9b19)), closes [#513](https://github.com/rehearsal-js/rehearsal-js/issues/513)
+* fix: reset files when cli falis (#520) ([bfcbd9a](https://github.com/rehearsal-js/rehearsal-js/commit/bfcbd9a)), closes [#520](https://github.com/rehearsal-js/rehearsal-js/issues/520)
+* fix: terminal based editor in interactive mode (#594) ([1733e10](https://github.com/rehearsal-js/rehearsal-js/commit/1733e10)), closes [#594](https://github.com/rehearsal-js/rehearsal-js/issues/594)
+* fix: use relative paths in report (#540) ([905f40e](https://github.com/rehearsal-js/rehearsal-js/commit/905f40e)), closes [#540](https://github.com/rehearsal-js/rehearsal-js/issues/540)
+* fix: refactor super for packages (#597) ([c939f5e](https://github.com/rehearsal-js/rehearsal-js/commit/c939f5e)), closes [#597](https://github.com/rehearsal-js/rehearsal-js/issues/597)
+* feat: accept/edit/discard file prompt in interactive mode (#558) ([dc2c11e](https://github.com/rehearsal-js/rehearsal-js/commit/dc2c11e)), closes [#558](https://github.com/rehearsal-js/rehearsal-js/issues/558)
+* feat: add dry run option in migrate CLI (#608) ([ad62da4](https://github.com/rehearsal-js/rehearsal-js/commit/ad62da4)), closes [#608](https://github.com/rehearsal-js/rehearsal-js/issues/608)
+* feat: add eslint report, report any (#574) ([a608325](https://github.com/rehearsal-js/rehearsal-js/commit/a608325)), closes [#574](https://github.com/rehearsal-js/rehearsal-js/issues/574)
+* feat: ignore css imports (#550) ([65ebca0](https://github.com/rehearsal-js/rehearsal-js/commit/65ebca0)), closes [#550](https://github.com/rehearsal-js/rehearsal-js/issues/550)
+* feat: include exclude patterns from module graph (#588) ([61d32fe](https://github.com/rehearsal-js/rehearsal-js/commit/61d32fe)), closes [#588](https://github.com/rehearsal-js/rehearsal-js/issues/588)
+* feat: include/exclude patterns; add package scope (#575) ([593ba7c](https://github.com/rehearsal-js/rehearsal-js/commit/593ba7c)), closes [#575](https://github.com/rehearsal-js/rehearsal-js/issues/575)
+* feat: install @types/node (#611) ([1f1b5f9](https://github.com/rehearsal-js/rehearsal-js/commit/1f1b5f9)), closes [#611](https://github.com/rehearsal-js/rehearsal-js/issues/611)
+* feat: more correct formatting (#557) ([2f9ffd8](https://github.com/rehearsal-js/rehearsal-js/commit/2f9ffd8)), closes [#557](https://github.com/rehearsal-js/rehearsal-js/issues/557)
+* feat: prioritize fixing shape (#569) ([1d41705](https://github.com/rehearsal-js/rehearsal-js/commit/1d41705)), closes [#569](https://github.com/rehearsal-js/rehearsal-js/issues/569)
+* feat: root package and edges (#589) ([bb7d764](https://github.com/rehearsal-js/rehearsal-js/commit/bb7d764)), closes [#589](https://github.com/rehearsal-js/rehearsal-js/issues/589)
+* feat: use esModuleInterop (#548) ([c91cc44](https://github.com/rehearsal-js/rehearsal-js/commit/c91cc44)), closes [#548](https://github.com/rehearsal-js/rehearsal-js/issues/548)
+* feat: add support for workspaces for a js library (#515) ([e35e174](https://github.com/rehearsal-js/rehearsal-js/commit/e35e174)), closes [#515](https://github.com/rehearsal-js/rehearsal-js/issues/515)
+* chore: import type module (#571) ([fd1ac8e](https://github.com/rehearsal-js/rehearsal-js/commit/fd1ac8e)), closes [#571](https://github.com/rehearsal-js/rehearsal-js/issues/571)
+* chore: project references (#512) ([e946c54](https://github.com/rehearsal-js/rehearsal-js/commit/e946c54)), closes [#512](https://github.com/rehearsal-js/rehearsal-js/issues/512)
+* chore: remove baseurl (#565) ([97fc972](https://github.com/rehearsal-js/rehearsal-js/commit/97fc972)), closes [#565](https://github.com/rehearsal-js/rehearsal-js/issues/565)
+* chore: remove unused babel dependencies (#488) ([cacd829](https://github.com/rehearsal-js/rehearsal-js/commit/cacd829)), closes [#488](https://github.com/rehearsal-js/rehearsal-js/issues/488)
+* chore: removed unused dependencies (#551) ([1ee0059](https://github.com/rehearsal-js/rehearsal-js/commit/1ee0059)), closes [#551](https://github.com/rehearsal-js/rehearsal-js/issues/551)
+* chore: test for 'diagnostic ordering' commit (#546) ([e9ef233](https://github.com/rehearsal-js/rehearsal-js/commit/e9ef233)), closes [#546](https://github.com/rehearsal-js/rehearsal-js/issues/546)
+
 ##  v0.0.39
 
 * fix: excessive inline comments (#478) ([000592d](https://github.com/rehearsal-js/rehearsal-js/commit/000592d)), closes [#460](https://github.com/rehearsal-js/rehearsal-js/issues/478)
@@ -13,7 +96,7 @@
 * feat: improve service discovery util (#420) ([fdb4cf3](https://github.com/rehearsal-js/rehearsal-js/commit/fdb4cf3)), closes [#420](https://github.com/rehearsal-js/rehearsal-js/issues/420)
 * feat: keep git history during converting JS to TS files (#422) ([55de74d](https://github.com/rehearsal-js/rehearsal-js/commit/55de74d)), closes [#422](https://github.com/rehearsal-js/rehearsal-js/issues/422)
 * feat: migration state (#440) ([99d9ef5](https://github.com/rehearsal-js/rehearsal-js/commit/99d9ef5)), closes [#440](https://github.com/rehearsal-js/rehearsal-js/issues/440)
-* feat: minumum interactive mode (#410) ([a45536a](https://github.com/rehearsal-js/rehearsal-js/commit/a45536a)), closes [#410](https://github.com/rehearsal-js/rehearsal-js/issues/410)
+* feat: minimum interactive mode (#410) ([a45536a](https://github.com/rehearsal-js/rehearsal-js/commit/a45536a)), closes [#410](https://github.com/rehearsal-js/rehearsal-js/issues/410)
 * feat: tighten default tsconfig (#452) ([f56c61b](https://github.com/rehearsal-js/rehearsal-js/commit/f56c61b)), closes [#452](https://github.com/rehearsal-js/rehearsal-js/issues/452)
 * feat: use typescript codefixes with type inference (#459) ([74db741](https://github.com/rehearsal-js/rehearsal-js/commit/74db741)), closes [#459](https://github.com/rehearsal-js/rehearsal-js/issues/459)
 * feat(tests): create test-support package (#427) ([2c52484](https://github.com/rehearsal-js/rehearsal-js/commit/2c52484)), closes [#427](https://github.com/rehearsal-js/rehearsal-js/issues/427)